\name{lmBuffer-package}
\alias{lmBuffer-package}
\alias{lmBuffer}
\docType{package}
\title{
\packageTitle{lmBuffer}
}
\description{
\packageDescription{lmBuffer}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lmBuffer}
\packageIndices{lmBuffer}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{lmBuffer}

Maintainer: \packageMaintainer{lmBuffer}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
