% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overList.R
\name{overList}
\alias{overList}
\title{Array index to a list}
\usage{
overList(overBuff, nPlot)
}
\arguments{
\item{overBuff}{Index from an array (distance matrix) created with 'which' argument arr.index}

\item{nPlot}{Maximum index number}
}
\value{
A list with  of \code{x} and \code{y}
}
\description{
Array index to a list
}
\examples{
distVF <- as.matrix(dist(1:10)) > 5
indArr <- which(distVF, arr.ind=TRUE)
overList(indArr, 10)
}

